requirements.txt
opencv-python
scenedetect
pillow
yt-dlp
